+++ lightstep/sdk/metric/internal/pipeline/overflow.go
@@ -1,4 +1,4 @@
-// Copyright Lightstep Authors
+// Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,6 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package launcher
+package pipeline // import "github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/internal/pipeline"
 
-const version = "1.14.0"
+import "go.opentelemetry.io/otel/attribute"
+
+// OverflowAttributes is the specified list of attributes to use when
+// configured mechanisms overflow a cardinality limit.
+var OverflowAttributes = []attribute.KeyValue{
+	attribute.Bool("otel.metric.overflow", true),
+}
+
+// OverflowAttributeSet is the set corresponding with OverflowAttributes.
+var OverflowAttributeSet = attribute.NewSet(OverflowAttributes...)